v2_2025-01/code/requirements.txt
openai
google-generativeai
anthropic
numpy
matplotlib
scikit-learn